Fix bug in JSON base64 encoding
Previously there was a bug in the base64 encoding implementation used by
`msgspec.json` for encoding `bytes`/`bytearray`/`memoryview` objects.
This bug only appeared for bytes with values outside the range 0 <= x <=
127, which weren't caught by the existing tests.

To remedy this we:
- fix the bug
- augment the existing test suite to cover this situation
- add some fuzz testing for the base64 encoding code
